Fix inadvertently skipping over unsupported-objects
In case IRR Explorer receives a "DEL xxx" for an unsupported object,
such as a domain or inetnum, followed by an "ADD xxx" for a route
object (which we do support!), IRR Explorer would delete the
route-object because the first DEL was still in effect. Software would
keep searching for a recognisable object ignoring any intermediate
object-gaps.